feat: hello world for egg v4
part of https://github.com/eggjs/egg/issues/3644
