+++ helloworld/.eslintrc
@@ -1,0 +1,6 @@
+{
+  "extends": [
+    "eslint-config-egg/typescript",
+    "eslint-config-egg/lib/rules/enforce-node-prefix"
+  ]
+}
