+++ helloworld/app/router.ts
@@ -1,6 +1,6 @@
-'use strict';
+import { Application } from 'egg';
 
-module.exports = app => {
+export default (app: Application) => {
   app.router.get('/', app.controller.home.render);
   app.router.get('/foo', app.controller.foo.render);
 };
