--- helloworld/config/config.default.js
@@ -1,3 +1,0 @@
-'use strict';
-
-exports.keys = 'my secret keys';
