+++ helloworld/config/config.default.ts
@@ -1,0 +1,3 @@
+export default {
+  keys: 'my secret keys',
+};
