+++ helloworld/package.json
@@ -1,16 +1,29 @@
 {
   "name": "helloworld",
   "dependencies": {
-    "egg": "^1.10.1"
+    "egg": "beta"
   },
   "devDependencies": {
-    "egg-bin": "^4.3.5",
-    "egg-mock": "^3.13.1"
+    "@eggjs/bin": "7",
+    "@eggjs/mock": "6",
+    "@eggjs/tsconfig": "1",
+    "@types/mocha": "10",
+    "@types/node": "22",
+    "eslint": "8",
+    "eslint-config-egg": "14",
+    "typescript": "5"
   },
   "scripts": {
+    "lint": "eslint . --ext .ts",
     "dev": "egg-bin dev",
+    "pretest": "npm run lint -- --fix",
     "test": "egg-bin test",
-    "cov": "egg-bin cov"
+    "preci": "npm run lint",
+    "ci": "egg-bin cov",
+    "postci": "npm run prepublishOnly && npm run clean",
+    "clean": "tsc -b --clean",
+    "prepublishOnly": "npm run clean && tsc"
   },
-  "private": true
+  "private": true,
+  "repository": "git@github.com:eggjs/examples.git"
 }
